feat: code-refactor added image carousal in banner
